move files around
simplify code




b


fixes to ising


fixes for jax impl
